+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changed
+- Cleanup private key only for file or pem based certificate login (#2101)
+
 ## [3.10.1906.0 - Unreleased]
 
 ### Added