+++ CHANGELOG.md
@@ -5,9 +5,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+### Changed
+
+- Cleanup private key only for file or pem based certificate login (#2101)
+- ConvertTo-PnPClientSidePage: Added support to transform web part pages that live outside of a library (so in the root of the site)
+
 ## [3.9.1905.3 - May 2019 Intermediate Release 3]
 
 ### Changed
+
 - Updated core provisioning to handle token issue during extraction and reintroduced content type fieldlink reordering in the engine.
 
 ## [3.9.1905.2 - May 2019 Intermediate Release 2]